custom_pc_odoo_17_v6/models/pc_builder.py
```python
from odoo import models, fields, api

class PCBuilder(models.Model):
    _name = 'pc.builder'
    _description = 'PC Builder'

    name = fields.Char(string='Name', required=True)
    cpu_id = fields.Many2one('product.product', string='CPU')
    gpu_id = fields.Many2one('product.product', string='GPU')
    motherboard_id = fields.Many2one('product.product', string='Motherboard')
    ram_id = fields.Many2one('product.product', string='RAM')
    storage_id = fields.Many2one('product.product', string='Storage')
    power_supply_id = fields.Many2one('product.product', string='Power Supply')
    cooling_solution_id = fields.Many2one('product.product', string='Cooling Solution')
    case_id = fields.Many2one('product.product', string='Case')

    @api.onchange('cpu_id', 'gpu_id', 'motherboard_id', 'ram_id', 'storage_id', 'power_supply_id', 'cooling_solution_id', 'case_id')
    def _check_compatibility(self):
        # Implement compatibility logic here
        pass

    @api.model
    def create(self, vals):
        # Implement additional logic during creation of a PC build
        return super(PCBuilder, self).create(vals)

    def write(self, vals):
        # Implement additional logic during update of a PC build
        return super(PCBuilder, self).write(vals)

    def unlink(self):
        # Implement additional logic during deletion of a PC build
        return super(PCBuilder, self).unlink()
```